requirements.txt
langchain
langchain_openai
langchain_azure_openai
openai
huggingface_hub
python-dotenv
streamlit